refactor(editor): 에디터 화면 템플릿마다 path 달라지도록 변경
이전에는 editor 라우트에 모든 템플릿이 보였다면 이제는 editor/1 형식으로 뒤에 템플릿 아이디가 붙게됨
