+++ card-capture/src/app/editor/[templateId]/layout.tsx
