+++ card-capture/src/app/editor/[templateId]/page.tsx
@@ -1,6 +1,6 @@
 import dynamic from 'next/dynamic';
 
-const EditorPage = dynamic(() => import('./EditorContent'), {
+const EditorPage = dynamic(() => import('@/app/editor/[templateId]/EditorContent'), {
   ssr: false,
   loading: () => <p></p>,
 });
